refactor: data 패키지 분리
