+++ android/app/src/main/java/io/coursepick/coursepick/data/Services.kt
@@ -2,6 +2,10 @@ package io.coursepick.coursepick.data
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import io.coursepick.coursepick.BuildConfig
+import io.coursepick.coursepick.data.course.CourseService
+import io.coursepick.coursepick.data.interceptor.KakaoAuthInterceptor
+import io.coursepick.coursepick.data.interceptor.PrettyPrintLogger
+import io.coursepick.coursepick.data.search.SearchService
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
