+++ android/app/src/main/java/io/coursepick/coursepick/data/course/CourseDto.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.data
+package io.coursepick.coursepick.data.course
 
 import io.coursepick.coursepick.domain.course.Course
 import io.coursepick.coursepick.domain.course.CourseName
