+++ android/app/src/main/java/io/coursepick/coursepick/data/course/DefaultCourseRepository.kt
@@ -1,5 +1,6 @@
-package io.coursepick.coursepick.data
+package io.coursepick.coursepick.data.course
 
+import io.coursepick.coursepick.data.Services
 import io.coursepick.coursepick.domain.course.Coordinate
 import io.coursepick.coursepick.domain.course.Course
 import io.coursepick.coursepick.domain.course.CourseRepository
