+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CoursesViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import io.coursepick.coursepick.data.DefaultCourseRepository
+import io.coursepick.coursepick.data.course.DefaultCourseRepository
 import io.coursepick.coursepick.domain.course.Coordinate
 import io.coursepick.coursepick.domain.course.Course
 import io.coursepick.coursepick.domain.course.CourseRepository
